fix(csharp/src/Drivers/Databricks): Set enable_run_async_thrift default true (#3232)
## Motivation

In PR https://github.com/apache/arrow-adbc/pull/3171, `RunAsync` option
in `TExecuteStatementReq` was added and exposed via connection parameter
`adbc.databricks.enable_run_async_thrift`, but it is not enabled by
default. This is turned on by default in other Databricks drivers, we
should turn in on by default in ADBC as well.

## Change
- Set `DatabricksConnection:_runAsyncInThrift` default value to `true`

## Test
- PBI PQTest with all the test cases