fix(csharp/src/Drivers/Databricks): update error type for connection errors when possible (#3581)
The driver currently throws an AggregateException when an error occurs
during the `connection.OpenAsync().Wait();` call. This PR attempts to
unwind the AggregateException and throw an AdbcException to conform to
the ADBC spec. It keeps the AggregateException as the InnerException of
the throw AdbcException to maintain the integrity of the stack.

Co-authored-by: David Coe <>