feat(go): Add Databrikcs driver written in Go (#3325)
This PR is a continuation of
https://github.com/apache/arrow-adbc/pull/2998

Key changes from Jade's PR:
- sql.DB has been pushed up to Database as a stored connection pool
- Connections are no longer pooled on adbc Connections - the raw
connection is persisted
- Bind and BindStream are marked as todo rather than partial
implementations
- Redundant tests have been cleaned up
- Reference counting was implemented on the custom IPC reader which was
causing readers to be prematurely destroyed before consumption

**Connection Pooling**
Databricks offers a `sql.DB` struct which manages a connection pool.
This is now initialized on the ADBC `Database` struct when connections
are opened and re-used if no options have changed on the Database.

Connections can be obtained from the pool which are stored on the ADBC
Connection.

---------

Co-authored-by: Jade Wang <jade.wang@databricks.com>